+++ src/main/java/org/nrg/containers/model/command/entity/DockerSetupCommandEntity.java
@@ -1,8 +1,5 @@
 package org.nrg.containers.model.command.entity;
 
-import org.hibernate.envers.Audited;
-import org.nrg.containers.model.command.auto.Command;
-
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
 import javax.persistence.Transient;
@@ -11,10 +8,6 @@ import javax.persistence.Transient;
 @DiscriminatorValue("docker-setup")
 public class DockerSetupCommandEntity extends CommandEntity {
     public static final CommandType type = CommandType.DOCKER_SETUP;
-
-    public static DockerSetupCommandEntity fromPojo(final Command commandPojo) {
-        return new DockerSetupCommandEntity();
-    }
 
     @Transient
     public CommandType getType() {